input/loader-producer/requirements.txt
aiohttp
python-dotenv
aiokafka
uvloop
orjson
loguru
